Complete Backend Auth API & Installed React Frontend

+++ backend/controllers/carController.js
@@ -1,12 +1,14 @@
 // using express-async-handler package to avoid using the try{}catch{} syntax
 const asyncHandler = require('express-async-handler');
 const Car = require('../models/carModel');
+// used for put and delete requests only
+const User = require('../models/userModel');
 
 // @desc Get cars
 // @route GET /api/cars
 // @access Private
 const getCars = asyncHandler(async (req, res) => {
-  const cars = await Car.find();
+  const cars = await Car.find({ user: req.user.id });
   res.status(200).json(cars);
 });
 
@@ -23,6 +25,7 @@ const postCar = asyncHandler(async (req, res) => {
   // send to DB is text is sent
   const car = await Car.create({
     text: req.body.text,
+    user: req.user.id,
   });
   res.status(200).json(car);
 });
@@ -39,6 +42,21 @@ const updateCar = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error('Car not found');
   }
+
+  // check for user from the User model
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    // throw error if user is not found
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // check logged in user matches car user
+  if (car.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
+  }
+
   // update DB if id is found. => id/textData/options[create if doesn't already exist]
   const updatedCar = await Car.findByIdAndUpdate(id, req.body, { new: true });
   res.status(200).json(updatedCar);
@@ -55,6 +73,20 @@ const deleteCar = asyncHandler(async (req, res) => {
   if (!car) {
     res.status(400);
     throw new Error('Car not found');
+  }
+
+  // check for user from the User model
+  const user = await User.findById(req.user.id);
+  if (!user) {
+    // throw error if user is not found
+    res.status(401);
+    throw new Error('User not found');
+  }
+
+  // check logged in user matches car user
+  if (car.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error('User not authorized');
   }
   // delete DB if id is found
   const deletedCar = await Car.findByIdAndDelete(id);
